4.0.1/dark mode (#5)
* Updated Font Awesome, ESLint, Gatsby, and React dependencies to their
latest version.

* Updated the "Reset" page to correct wording under the "Heading classes"
section. The text originally said "...in the event that an actual heading
element does make sense...", whereas the original intention was that
heading classes should be used where actual heading elements do not make
sense.

TL;DR - Changed does to doesn't in paragraph under the "Heading classes"
section of the "Reset" page.

* Updated the following dependencies to their latest version:

- eslint
- gatsby
- gatsby-image
- gatsby-plugin-manifest
- gatsby-plugin-offline
- gatsby-plugin-sharp
- gatsby-source-filesystem
- gatsby-transformer-sharp
- node-sass
- react
- react-dom

* Updated the shared media query SASS partial to rename the "media-query"
mixin to "breakpoint", as there's a new media query mixin titled
"color-scheme", so differentiating the two was critical for code
readability and maintainability.

* Updated all SASS component and page files to respect system theme
(prefers-color-scheme) automatically using CSS variables. If a user has
their system theme set to "dark", you can expect:

1. The background color to be set to black with text in white (perfect
for OLED screens)
2. Perks, code blocks, and UI borders set to a dark gray to still stand
out from the background
3. Muted accent colors (the header gradient and scroll-to-top button is darkened by 10%, and
alert colors are darkened by 5%)

* Updated the package.json to include a shorthand for the development
script, e.g. "dev"

* Updated the footer React component to be a proper class with an added
button in the footer__links section that changes the theme by adding a
theme data-attribute to the document. In future updates, overriding the
system theme will create a cookie to persist the override while
navigating the site or upon later visitation.

* Updated all component and page SASS files to not only include light/dark
themes for system theme preferences (prefers-color-scheme), but for
theme overrides use data-attributes on the document element. This can be
triggered using the new theme changing button in the footer's
footer__links section.

In addition, the footer component's SASS file was updated to properly
style the theme changing button.

* Added a new dependency, js-cookie. This will be used for storing theme
data in the event that a user specifies a different theme than that of
the system's default.

* Updated the footer component to remember theme overrides, allowing users
to navigate the site without the site theme reverting to the system
default. This system is also smart, and if an override later matches
that of the system default, all saved data will be removed; This allows
for the theme to dynamically change as the user changes the system
theme.