4.0.1/dark mode fix (#6)
* Updated the following dependencies to their latest versions:

- gatsby
- gatsby-image
- gatsby-plugin-manifest
- gatsby-plugin-offline
- gatsby-plugin-react-helmet
- gatsby-plugin-sass
- gatsby-plugin-sharp
- gatsby-plugin-sitemap
- gatsby-source-filesystem
- gatsby-transformer-sharp
- react-copy-to-clipboard
- prettier

* Updated the footer React component to create a theme override cookie
that expires in a year as opposed to at the end of the session. This
way, if the user closes the browser and navigates to the site at a later
date, it will be with the theme the user has selected.

* Updated the header component SASS to change the highlighted button in
the header navigation when in dark mode. This follows the styling
similar to that of the index page's button, and will hopefully make the
site more cohesive when in dark mode.

* Updated the readme.md to include documentation explaining that the
shorthand "npm run dev" can be used, as this was implemented in the last
update but was not documented properly for those pulling down the site.