Code Component Update (#12)
Updated the Code React component to no longer dangerously set the HTML of the example with the child string. Instead, it creates the wrapping DIV with a reference which later has its innerHTML set once the component mounts. This cleans up the code a bit and makes it far more readable.