New card component
Since I wanted to use the same cards found on the index page on the 404
page to better compartmentalize information, I created a new Card
component with its own styles and structure. This means that the index
page was updated to use this new component, with its styles reduced as it
no longer needs to include card styling, and the 404 page now includes
this component.