4.0.5 (#14)
* Dependency updates

Updated `chassis-css` to v4.0.5. In addition, the following dependencies
have been updated to their latest versions:

- eslint@7.4.0
- gatsby-image@2.4.13
- gatsby-plugin-offline@3.2.18
- eslint-plugin-react@7.20.3
- gatsby@2.24.3
- @fortawesome/free-brands-svg-icons@5.14.0
- @fortawesome/react-fontawesome@0.1.11
- gatsby-plugin-react-helmet@3.3.10
- gatsby-plugin-sass@2.3.12
- gatsby-plugin-manifest@2.4.18
- @fortawesome/fontawesome-svg-core@1.2.30
- @fortawesome/free-solid-svg-icons@5.14.0
- node-sass@4.14.1
- gatsby-source-filesystem@2.3.19
- prettier@2.0.5
- react-helmet@6.1.0
- gatsby-plugin-sitemap@2.4.11
- gatsby-plugin-sharp@2.6.19
- gatsby-transformer-sharp@2.5.11

* 4.0.5 release

Updated the `gatsby-config.js` and index page to include the latest
release's version number and summary respectively.

* New card component

Since I wanted to use the same cards found on the index page on the 404
page to better compartmentalize information, I created a new Card
component with its own styles and structure. This means that the index
page was updated to use this new component, with its styles reduced as it
no longer needs to include card styling, and the 404 page now includes
this component.

* Dark mode links brightened

Updated the dark mode styling of links to lighten their regular/hover
colors to look more readable.

* Lodash security vulnerability fix

Ran an `npm update lodash --depth 10` to fix the 57 associated security
vulnerability warnings.

* Build warning fixes

Updated the Card component's styles to remove unused imports, as well as
the index page's to reorder the component imports; These both were to
fix `mini-css-extract-plugin` "Conflicting order" errors on build.