4.0.8 (#18)
**Version bump**

Bumped the version of the package to v4.0.8.

**Node.js bump**

Bumped the Node.js version to v16.13.0.

**Dependency updates & Gatsby 4**

Updated all outdated dependencies to their latest versions. Additionally, all Gatsby packages have been upgraded to their latest, Gatsby 4 versions.

**Chassis.css bump**

Bumped the version of `chassis-css` to v4.0.8.

**Browserslist changes**

Updated the `browserslist` in the `package.json` to convert it to an array and to drop OperaMini support, as it's impossible to target JS for the browser due to its different performance modes.

**A11Y improvements & general updates**

Updated many of the site components to fix WCAG AA 2.1 A11Y issues as well as provide up-to-date information regarding the v4.0.8 patch release of Chassis.css.